Refactored the code as per feedback. Now all code is .h file located in src folder, the examples folder contains 2 examples .ino files
