+++ library.properties
@@ -1,0 +1,10 @@
+name=Qbead
+version=1.0.0
+author= 
+maintainer= 
+sentence=A library for controlling the Qbead device.
+paragraph=The Qbead library provides functionalities to control the Qbead device, including state management, LED control, and IMU interaction.
+url= 
+architectures=nrf52
+includes=Qbead.h
+depends=Adafruit NeoPixel, Seeed Arduino LSM6DS3
